// src/enums/categories.ts
enum categories {
    // Software
    WEB_DEVELOPMENT = 'Web Development',
    APP_DEVELOPMENT = 'App Development',
    DESKTOP_SOFTWARE_DEVELOPMENT = 'Desktop Software Development',
    ECOMMERCE_DEVELOPMENT = 'Ecommerce Development',
    GAME_DEVELOPMENT = 'Game Development',
    TRADING_SOFTWARE_DEVELOPMENT = 'Trading Software Development',
    BLOCKCHAIN_DEVELOPMENT = 'Blockchain Development',
    MACHINE_LEARNING = 'Machine Learning',
    ARTIFICAL_INTELLIGENCE = 'Artificial Intelligence',
    SOFTWARE_TESTING = 'Software Testing',

    // Design
    WEB_DESING = 'Web Design',
    MOBILE_DESIGN = 'Mobile Design',
    GAME_DESIGN = 'Game Design',
    GRAPHIC_DESIGN = 'Graphic Design',
    BRAND_DESIGN = 'Brand Design',
    PRODUCT_DESIGN = 'Product Design',

    // Media
    ANIMATION = 'Animation',
    ILLUSTRATION = 'Illustration',
    PHOTOGRAPHY = 'Photography',
    WRITING = 'Writing',
    CONTENT_RESEARCH = 'Content Research',
    AUDIO_PRODUCTION = 'Audio Production',
    VIDEO_PRODUCTION = 'Video Production',

    // Business
    BUSINESS_PROJECT_MANAGEMENT = 'Business Project Management',
    CUSTOMER_SERVICE = 'Customer Service',
    TECHNICAL_SUPPORT = 'Technical Support',
    ADVERTISING = 'Advertising',
    MARKETING = 'Marketing',
    ACCOUNTING = 'Accounting',
    BUSINESS_RESEARCH = 'Business Research',
    LEGAL_RESOURCES = 'Legal Resources',
    HUMAN_RESOURCES = 'Human Resources',
    RECRUITING = 'Recruiting',

    OTHER = 'Other',
}

export default categories